[SPARK-44227][SQL] Extract SchemaUtils from StructField
### What changes were proposed in this pull request?

Extract SchemaUtils from StructField so that StructField does not requires anything from Catalyst.

### Why are the changes needed?

To help StructField get rid of the need to depend on Catalyst.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing tests

Closes #41772 from amaliujia/move_schema_utils.

Authored-by: Rui Wang <rui.wang@databricks.com>
